requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.14.0

# Data processing and analysis
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Web UI framework
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Database
sqlite3  # Built-in with Python

# Additional utilities
dataclasses-json>=0.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For better performance
accelerate>=0.24.0
bitsandbytes>=0.41.0  # For 8-bit quantization
